Add notes for Day 2 of Elm to 7ml7w
We went off book for this day because the book covers features of Elm that
have been removed in the latest and the book relies on using online tools
to write the examples.  This makes it hard to actually play along.  However
I added notes about both the book, and the alternate reading material because
I'm some kind of masochist?