Notes for Elm day 3
I was unable to attend the meetings, and am playing catch up with what the group
covered.  The group didn't use the book for the 2nd and 3rd meetings on Elm and
we chose to use "The Elm Architecture" instead.  In Day 2 we didn't get through
that whole article so my notes from Day 2 cover what the group did on Day 3 too.
I did read the rest of the chapter though and my brief notes on that are what
I'm adding here.