Add Elixir Day 3 notes
Once more, I have not done the exercises so this is just my notes and the
implementation required by the chapter.  The book suggests adding code to a
day3 folder, but in order to get it to work I had to pull in most of the code
from day 2 as well.  I couldn't be bothered working out a sensible way to share
things, so this is mostly duplication.