Add miniKanren notes
We decided to do the whole chapter in one go but I've kept to the day_x
structure of the chapter for my notes.  There's also a Vagrantfile for
installing a clojure in a VM because I didn't want a JVM on my host.

There are notes for all 3 days and for days 1 and 2 at least I've filled in
some answers to some of the exercises.  For day 1 I've also included a
`scratchpad.clj` file that includes some of the repl session code I played
around with while reading.